Embedded annotation.
