+++ src/main/java/pl/abienkowska/todoapp/model/Audit.java
@@ -1,0 +1,24 @@
+package pl.abienkowska.todoapp.model;
+
+import javax.persistence.Embeddable;
+import javax.persistence.MappedSuperclass;
+import javax.persistence.PrePersist;
+import javax.persistence.PreUpdate;
+import java.time.LocalDateTime;
+
+@Embeddable
+class Audit {
+
+    private LocalDateTime createdOn;
+    private LocalDateTime updatedOn;
+
+    @PrePersist
+    void prePersist(){
+        createdOn = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    void preMerge(){
+        updatedOn = LocalDateTime.now();
+    }
+}
