+++ src/main/java/pl/abienkowska/todoapp/model/Task.java
@@ -15,8 +15,8 @@ public class Task {
     private String description;
     private boolean done;
     private LocalDateTime deadline;
-    private LocalDateTime createdOn;
-    private LocalDateTime updatedOn;
+    @Embedded
+    private Audit audit = new Audit();
 
     public Task(){
     }
@@ -44,15 +44,7 @@ public class Task {
         done = source.done;
         deadline = source.deadline;
     }
-    @PrePersist
-    void prePersist(){
-        createdOn = LocalDateTime.now();
-    }
 
-    @PreUpdate
-    void preMerge(){
-        updatedOn = LocalDateTime.now();
-    }
